Refactor network_utils imports
